Updates runOfflineTicks patch to exitOfflineLoop
In order to fix the removed `runOfflineTicks` patch that got removed,
this commit uses the `exitOfflineLoop` in its place. Additionally, it
updates the secondsSinceLoaded variable to be accessed a bit
differently, using Date.now() and the `this._offlineInfo` object for
offline time checking. Finally, it updates a minor typo in a comment.